# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gulden API",
    "version": "3.23.0",
    "description": "Static contract for the Gulden HTTP API. JSON amounts are numbers and transaction expenses are negative. Unless an operation is marked public, send the bp_session cookie returned by login or setup."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Gulden API base path"
    }
  ],
...
  "security": [{ "bpSession": [] }],
  "paths": {
...
    "/accounts": {
      "get": {
        "tags": ["Resources"],
        "summary": "List accounts",
        "operationId": "listAccounts",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create an account",
        "operationId": "createAccount",
        "requestBody": { "$ref": "#/components/requestBodies/Account" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/persons": {
      "get": {
        "tags": ["Resources"],
        "summary": "List people",
        "operationId": "listPersons",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create a person",
        "operationId": "createPerson",
        "requestBody": { "$ref": "#/components/requestBodies/Person" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/categories": {
      "get": {
        "tags": ["Resources"],
        "summary": "List categories",
        "operationId": "listCategories",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create a category",
        "operationId": "createCategory",
        "requestBody": { "$ref": "#/components/requestBodies/Category" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/transactions": {
      "get": {
        "tags": ["Resources"],
        "summary": "List transactions",
        "operationId": "listTransactions",
        "parameters": [
          { "$ref": "#/components/parameters/MonthQuery" },
          { "name": "review", "in": "query", "schema": { "type": "string", "enum": ["1"] } },
          { "name": "category_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "account_id", "in": "query", "schema": { "type": "integer" } },
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive substring match on the description.",
            "schema": { "type": "string" }
          },
          {
            "name": "tx_type",
            "in": "query",
            "description": "Exact case-insensitive match on tx_type.",
            "schema": { "type": "string" }
          },
          {
            "name": "date_from",
            "in": "query",
            "description": "Inclusive lower date bound (YYYY-MM-DD).",
            "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" }
          },
          {
            "name": "date_to",
            "in": "query",
            "description": "Inclusive upper date bound (YYYY-MM-DD).",
            "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" }
          },
          {
            "name": "amount_min",
            "in": "query",
            "description": "Inclusive lower bound on the absolute amount.",
            "schema": { "type": "number" }
          },
          {
            "name": "amount_max",
            "in": "query",
            "description": "Inclusive upper bound on the absolute amount.",
            "schema": { "type": "number" }
          },
          {
            "name": "transfer",
            "in": "query",
            "description": "1 = only paired transfers, 0 = exclude them.",
            "schema": { "type": "string", "enum": ["0", "1"] }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Rows per page. The maximum is 1000; the special value \"all\" returns the whole filtered set (capped at 10000 rows) for the single-page table.",
            "schema": { "type": "string" }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of matching rows to skip.",
            "schema": { "type": "integer", "minimum": 0, "default": 0 }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create one or more manual transactions",
        "operationId": "createTransactions",
        "requestBody": { "$ref": "#/components/requestBodies/ManualTransactions" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/commitments": {
      "get": {
        "tags": ["Resources"],
        "summary": "List commitments",
        "operationId": "listCommitments",
        "parameters": [{ "$ref": "#/components/parameters/MonthQuery" }],
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create a commitment",
        "operationId": "createCommitment",
        "requestBody": { "$ref": "#/components/requestBodies/Commitment" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/funds": {
      "get": {
        "tags": ["Resources"],
        "summary": "List funds and movements",
        "operationId": "listFunds",
        "parameters": [{ "$ref": "#/components/parameters/MonthQuery" }],
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create a fund",
        "operationId": "createFund",
        "requestBody": { "$ref": "#/components/requestBodies/Fund" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/income": {
      "get": {
        "tags": ["Resources"],
        "summary": "Get income sources for a month",
        "operationId": "getIncome",
        "parameters": [{ "$ref": "#/components/parameters/MonthQuery" }],
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      }
    },
...
    "/balances": {
      "get": {
        "tags": ["Resources"],
        "summary": "Get balance observations and reconciliation",
        "operationId": "listBalances",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Record an observed account balance",
        "operationId": "createBalanceObservation",
        "requestBody": { "$ref": "#/components/requestBodies/BalanceObservation" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" },
          "404": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/projection": {
      "get": {
        "tags": ["Reports"],
        "summary": "Project future monthly balances",
        "operationId": "getProjection",
        "parameters": [
          {
            "name": "months",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 240, "default": 96 }
          },
          { "name": "from", "in": "query", "schema": { "$ref": "#/components/schemas/Month" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      }
    },
...
    "/shares": {
      "get": {
        "tags": ["Settings"],
        "summary": "List read-only budget share links",
        "operationId": "listShares",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Create a read-only budget share link",
        "operationId": "createShare",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["month"],
                "properties": {
                  "month": { "$ref": "#/components/schemas/Month" },
                  "expires_in_days": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365,
                    "default": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/recurrences": {
      "get": {
        "tags": ["Resources"],
        "summary": "List recurring transactions",
        "operationId": "listRecurrences",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Create a recurring transaction",
        "operationId": "createRecurrence",
        "requestBody": { "$ref": "#/components/requestBodies/Recurrence" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/attachments": {
      "get": {
        "tags": ["Resources"],
        "summary": "List attachments for a transaction",
        "operationId": "listAttachments",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      },
      "post": {
        "tags": ["Resources"],
        "summary": "Attach a file to a transaction",
        "operationId": "createAttachment",
        "requestBody": { "$ref": "#/components/requestBodies/AttachmentUpload" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" },
          "413": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
    "/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get user settings and available AI profiles",
        "operationId": "getSettings",
        "responses": { "200": { "$ref": "#/components/responses/JsonResponse" } }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update user settings",
        "operationId": "updateSettings",
        "requestBody": { "$ref": "#/components/requestBodies/Settings" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorResponse" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bpSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bp_session",
        "description": "HttpOnly session cookie issued by /auth/login or /auth/setup."
      }
    },
    "parameters": {
...
      "MonthQuery": {
        "name": "month",
        "in": "query",
        "schema": { "$ref": "#/components/schemas/Month" }
      }
    },
    "responses": {
...
      "JsonResponse": {
        "description": "Endpoint-specific JSON response.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Json" } } }
      },
      "ErrorResponse": {
        "description": "A failed request.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "requestBodies": {
...
      "Account": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Account" } } }
      },
...
      "Person": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Person" } } }
      },
      "Category": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } }
      },
...
      "ManualTransactions": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                { "$ref": "#/components/schemas/ManualTransaction" },
                {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ManualTransaction" },
                  "maxItems": 200
                }
              ]
            }
          }
        }
      },
...
      "Commitment": {
        "required": true,
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Commitment" } }
        }
      },
...
      "Fund": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Fund" } } }
      },
...
      "BalanceObservation": {
        "required": true,
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/BalanceObservation" } }
        }
      },
      "Recurrence": {
        "required": true,
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Recurrence" } }
        }
      },
...
      "AttachmentUpload": {
        "required": true,
        "content": {
          "multipart/form-data": {
            "schema": { "$ref": "#/components/schemas/AttachmentMultipart" }
          }
        }
      },
      "Settings": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Settings" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } },
        "additionalProperties": true
      },
      "Json": {
        "description": "Endpoint-specific JSON object or array. The API intentionally returns different shapes per resource.",
        "oneOf": [
          { "type": "object", "additionalProperties": true },
          { "type": "array", "items": {} }
        ]
      },
      "Binary": { "type": "string", "format": "binary" },
...
      "AttachmentMultipart": {
        "type": "object",
        "required": ["file", "transaction_id"],
        "properties": {
          "file": { "$ref": "#/components/schemas/Binary" },
          "transaction_id": { "type": "integer", "minimum": 1 }
        }
      },
...
      "Month": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$", "example": "2026-08" },
...
      "Currency": { "type": "string", "pattern": "^[A-Z]{3}$", "default": "EUR" },
...
      "Account": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 60 },
          "kind": {
            "type": "string",
            "enum": ["bank", "card", "cash", "other", "sparkasse", "revolut"]
          },
          "is_spending_pot": { "type": "boolean" },
          "opening_balance": { "type": "number" },
          "opening_balance_month": {
            "type": "string",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "nullable": true,
            "description": "The opening balance was true at the END of this month; null covers all history"
          },
          "display_currency": { "type": "string", "enum": ["EUR", "USD", "GBP", "CHF"] }
        }
      },
...
      "Person": {
        "type": "object",
        "required": ["name"],
        "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 60 } }
      },
      "Category": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "group_id": { "type": ["integer", "null"] },
          "account_id": { "type": ["integer", "null"] },
          "monthly_budget": { "type": "number", "minimum": 0 },
          "active_from": { "type": ["string", "null"] },
          "active_to": { "type": ["string", "null"] },
          "is_active": { "type": "boolean" }
        }
      },
...
      "ManualTransaction": {
        "type": "object",
        "required": ["date", "description", "amount"],
        "properties": {
          "date": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "description": { "type": "string", "minLength": 1, "maxLength": 200 },
          "amount": { "type": "number", "not": { "const": 0 } },
          "currency": { "$ref": "#/components/schemas/Currency" },
          "tx_type": { "type": "string" },
          "account_id": { "type": ["integer", "null"] },
          "category_id": { "type": ["integer", "null"] },
          "fund_id": { "type": ["integer", "null"] },
          "commitment_id": { "type": ["integer", "null"] },
          "transfer_group": { "type": ["string", "null"] }
        }
      },
...
      "Commitment": {
        "type": "object",
        "required": ["name", "start_month"],
        "properties": {
          "name": { "type": "string" },
          "monthly_amount": { "type": "number", "minimum": 0 },
          "start_month": { "$ref": "#/components/schemas/Month" },
          "end_month": {
            "type": ["string", "null"],
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "description": "Inclusive end month; must not precede start_month."
          },
          "account_id": { "type": ["integer", "null"] },
          "fund_id": { "type": ["integer", "null"] },
          "category_id": { "type": ["integer", "null"] },
          "note": { "type": ["string", "null"] }
        }
      },
...
      "Fund": {
        "type": "object",
        "required": ["name", "start_month"],
        "properties": {
          "name": { "type": "string" },
          "monthly_contribution": { "type": "number", "minimum": 0 },
          "start_month": { "$ref": "#/components/schemas/Month" },
          "opening_balance": { "type": "number" },
          "category_id": { "type": ["integer", "null"] },
          "target_amount": { "type": ["number", "null"], "minimum": 0 },
          "target_date": { "type": ["string", "null"], "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" }
        }
      },
...
      "BalanceObservation": {
        "type": "object",
        "required": ["account_id", "month", "balance"],
        "properties": {
          "account_id": { "type": "integer" },
          "month": { "$ref": "#/components/schemas/Month" },
          "balance": { "type": "number" }
        }
      },
      "RecurrencePart": {
        "type": "object",
        "required": ["category_id", "amount"],
        "properties": {
          "category_id": { "type": "integer" },
          "amount": { "type": "number" }
        }
      },
      "Recurrence": {
        "type": "object",
        "required": ["name", "amount", "day_of_month"],
        "properties": {
          "name": { "type": "string" },
          "amount": { "type": "number" },
          "day_of_month": { "type": "integer", "minimum": 1, "maximum": 28 },
          "account_id": { "type": ["integer", "null"] },
          "category_id": { "type": ["integer", "null"] },
          "auto_post": { "type": "boolean" },
          "start_month": {
            "type": ["string", "null"],
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "description": "First scheduled month; null = scheduled from any month"
          },
          "end_month": {
            "type": ["string", "null"],
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "description": "Last scheduled month; null = ongoing"
          },
          "skip_months": {
            "type": "array",
            "items": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" },
            "description": "Months this item must NOT post or fold imported rows into"
          },
          "parts": {
            "type": "array",
            "minItems": 2,
            "items": { "$ref": "#/components/schemas/RecurrencePart" }
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "provider": { "type": "string" },
          "api_key": { "type": "string", "format": "password" },
          "model": { "type": "string" },
          "base_url": { "type": "string", "format": "uri" },
          "currency": { "type": "string", "enum": ["EUR", "USD", "GBP", "CHF"] },
          "theme": { "type": "string", "enum": ["system", "light", "dark", "midnight", "forest"] }
        }
      },
...
    }
  }
}